migrations: add suffix column to packages table
In order to eventually display all the package versions associated
with a specific series this new package suffix column will allow us
to join the 'packages' and 'modules' tables and fetch all the package
versions with the same package suffix and series name. These versions
will then be listed on the versions tab for the frontend.

Fixes b/129545571

Change-Id: I790187a0ea56cd238c7f88603995ea0fe883ba95
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/442289
Reviewed-by: Julie Qiu <julieqiu@google.com>