cmd: add environment variable GO_DISCOVERY_[service]_ADDR and default proxy to proxy.golang.org
The GO_DISCOVERY_[service]_ADDR variable is introduced, to set the adrr
that various services are running on. The default is set to
localhost:[port] to mute to popup on incoming networking connections
when running locally.

The default for GO_MODULE_PROXY_URL is set to https://proxy.golang.org
and for GO_MODULE_PROXY_INDEX is set to https://index.golang.org/index.

Change-Id: I097fd96a5c51ce1ec73ceab1928fda39de506352
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/445646
Reviewed-by: Robert Findley <rfindley@google.com>