frontend: refactor template parsing and fix active nav tab
This CL implements a pattern similar to that used in
github.com/golang/gddo to implement an active tab selector for the nav.
Along the way, refactor a bit so that the frontend package is fully
responsible for templates.

Also replace the pattern of Make*HandlerFunc with handler methods.

Also fix a minor typo in all.bash that was causing misspell not to be
installed.

Updates b/124309095

Change-Id: I802ba69b91b81be13c6314140326fe8783608a0c
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/445973
Reviewed-by: Julie Qiu <julieqiu@google.com>