cmd/cron,cmd/frontend: listen on port defined by environment variable PORT
App Engine requires that apps listen on the port defined by the PORT
environment variable. Otherwise, it logs an error message like the
following:

App is listening on port 8080, it should instead listen on the port
defined by the PORT environment variable. As a consequence, nginx cannot
be started. Performance may be degraded. Please listen on the port
defined by the PORT environment variable.

Change-Id: I38f8e0e96cc44e384fb712510efbfaaa68fef4f4
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/446246
Reviewed-by: Robert Findley <rfindley@google.com>