migrations: create search_results view
The search query is becoming difficult to read and maintain due to the
number of different tables that it involved.

A vw_search_results view is created, which contains all of the
information needed to generate search results.

postgres.Search is updated to use this view in CL 451331.

Updates b/130090305

Change-Id: I4149cd3e88e8b8b983578a56a9af294ea48d3af2
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/451330
Reviewed-by: Robert Findley <rfindley@google.com>