internal/postgres: update postgres.Search to use vw_search_results
The query for postgres.Search is updated to use vw_search_results, so
that the search query is easier to read and maintain.

Results with a rank less than 1e-10 are filtered out.

The following fields for postgres.SearchResult are changed:
* Relevance is deprecated
* Rank is added, which is calculated using relevance and num imported by
* NumImportedBy is now used instead of NumImporters, for consistency with
the frontend.
* Total is added, which represents the total number of results from the query

postgres.GetLatestPackageForPaths is deprecated, since it is no longer used.

A bug is fixed in details.tmpl where the active tab in the nav bar was not
rendering for importedby.

Updates b/130090305

Change-Id: I8d72c730bd8667b941ca7a519c23234bd6bd646d
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/451331
Reviewed-by: Robert Findley <rfindley@google.com>