discovery: increase the truthiness of our internal types
Break the reference cycle between 'Package' and 'Version', and get close
to the truth in the types returned from the database. To achieve this,
make the following changes:
 + Package type loses its Version reference
 + Version is split into metadata (VersionInfo) and Packages
 + VersionedPackage is introduced to capture the common case where we
   want both Package and Version metadata.

Additionally, ensure that Suffix is always set on Packages returned from
the database.

There are still a few fields that can 'lie', notably Package imports and
Version readme, but those are left to a potential later CL.

Updates b/130367504

Change-Id: I392238aca4c335b64766c2ef23d145f462d0ff1b
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/451770
Reviewed-by: Julie Qiu <julieqiu@google.com>