migrations: add a packages.redistributable column
In the event that a fetched package is detected as not redistributable,
we will store only approved metadata. In this case, I think it is
helpful to distinguish records in the database that have only partial
data.

A use case is the following: suppose in the future there is a change to
the set of metadata we can store for non-redistributable packages. This
column will allow us to easily select those packages that need
reprocessing.

Note that we will always be looking at licenses on the fly to determine
if a package is redistributable. This column serves the purpose of
storing whether or not we considered the package redistributable at the
time we fetched it.

Change-Id: Ibc23b0c80919749bd6bee6ee8b2ba6a5c0c99c46
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/455068
Reviewed-by: Julie Qiu <julieqiu@google.com>