content,internal/frontend: update imports, importedby and modules templates
The imports, importedby and module are updated to match the alpha designs.

Imports and ImportedBy now link to the latest version of each package and
display the package name.

A bug is fixed in HandleDetails, where an panic would occur if GetLatestPackage
returned an error and pkg.VersionInfo.Version was accessed immediately after.

Updates b/131593473

Change-Id: Icb42d1752f2d274cccc44cf5d47f994a3eaf2f62
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/455286
Reviewed-by: Robert Findley <rfindley@google.com>