internal/fetch: write an explicit go.mod file when it's missing
Some modules haven't had a go.mod file added to the repository yet.
The go tool uses an implicit module with module path equal to the
import path corresponding to the repository root, and no other
directives.

When trying to process such a module as the main module with the
go/packages library, there needs to be an explicit go.mod file for
go/packages to be able to find where the module is on disk. So,
write the implicit go.mod file explicitly. It's guaranteed not to
exist (since we write it only when it doesn't exist).

Add a test case to TestExtractPackagesFromZip and update some error
messages in it.

Fixes b/130040000

Change-Id: I3eb211aa67b5d87122ce74eaf081f34350e061be
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/456225
Reviewed-by: Julie Qiu <julieqiu@google.com>