cmd/fetch: replace deployment with statefulset and add PVCs
PersistentVolumes resources are added for writting to /tmp and /go.

Module zip files that are downloaded from the proxy are written to /tmp
for use by golang.org/x/tools/go/packages.Load for extracting package
information.

/go is set as the GOPATH in the Dockerfile.
golang.org/x/tools/go/packages.Load writes data to /go when running go
list.

Data stored in PersistentVolumes will continue to exist even as Pods are
deleted and recreated, which allows us to use them as a persistent
cache. As a result, the PersistentVolumes fetchdisk-tmp-pvc and
fetchdisk-gopath-pvc are created, with the ReadWriteOnce accessMode.

A StatefulSet is used to create the deployment, to prevent potential
deadlocks when pods are created and deleted. More info:
https://cloud.google.com/kubernetes-engine/docs/concepts/persistent-volumes#deployments_vs_statefulsets

cmd/fetch/README.md is updated with info on how to deploy the fetch
service with these changes.

Change-Id: Ic4964ec43c68edf20e1319c600158b824aff44c1
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/455000
Reviewed-by: Julie Qiu <julieqiu@google.com>