internal/postgres: recreate test database on migration failure
If switching between branches with different migrations, it's likely
that migration.Up will fail on your test databases. This can be a pain
now that there are multiple test databases to fix.

To handle this, automate the re-creation of the database if the
migration fails during the Up step.

Also, remove no-longer-needed migration for discovery-database-test.

Updates b/130719094

Change-Id: Ibb0550b95c2e7a3e437402cfe21ea22e460ac27b
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/457036
Reviewed-by: Julie Qiu <julieqiu@google.com>