internal/frontend: limit page number links on search page
At the moment, there is no limit on the number of page links on the
search page, which becomes messy when search has a lot of results.

Page numbers are now calculated using pagesToLink, which returns an
integer slice representing page numbers that will displayed. It
optimizes for the current page to be in the middle of that range
(similar to google.com).

Fixes b/131836875
Updates b/131862035

Change-Id: I9dafc35bff2278a4850bbded086a0000b5f76d3d
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/458088
Reviewed-by: Robert Findley <rfindley@google.com>