content,internal/frontend: separate imports and importedby by external/internal to module
Imports and importedby are now grouped based on whether they are internal or external
to the package's module. This makes it easier to understand information about
module dependencies.

Package names are also removed from the imports view to make it easier
to scan the package paths, and for consistency with the module and importedby
views.

Fixes b/132050335
Fixes b/132072505
Updates b/132071971

Change-Id: Ia849dd94129d2f5a95e56a43fd882e5801f0a532
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/460416
Reviewed-by: Robert Findley <rfindley@google.com>