content,internal/frontend: improve the message for hidden content
As pointed out in b/132284555, it's a little unclear what's going on
when a package is not redistributable: what exactly can't be shown?

This adds a 'DisplayName' property to details, so that we can improve
the error message to make it explicit that the details you're trying to
look at is intentionally hidden.

Fixes b/132284555

Change-Id: I25511ab15b317bb8ca0c53ef994b5f6d5b81a022
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/462576
Reviewed-by: Andrew Bonventre <andybons@google.com>