content,internal/frontend: leverage tab metadata to simplify details
As we've starting to parameterize tab metadata, it is now possible to
implement some simplifications to both our templates and our rendering.
In this CL, a formal 'TabSettings' type is introduced, and used to both
eliminate the contextual 'templateName' template func, and to generate
the module nav via a range action.

Change-Id: I454150158a9b9b5e401f45d52aeb2f00543d0cda
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/461525
Reviewed-by: Andrew Bonventre <andybons@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>