discovery: deprecate the fetch service in favor of appengine tasks
As a result of commit ca685af0e, we can now run the fetch action
entirely in-memory (and therefore, on AppEngine rather than K8s). This
opens up the possibility for a significant simplification of the fetch
process, which this CL (largely) implements. At a high level, this
change consists of:

 1. Move the fetch handler onto what is currently called the cron
 service.

 2. Introduce a task-queue abstraction for scheduling tasks that call
 back into the fetch endpoint. When running on GCP, this uses the
 AppEngine taskqueue package. When running locally, it uses a simple
 in-process queue.

 3. Replace the /indexupdate/ action with /poll-and-queue/, which both
 writes to the module_version_states table, and scheduled a fetch task.

This eliminates the need for the Kubernetes based fetch service, and
allows us to use Google Cloud Tasks for fetch tasks, which has some nice
advantages: built-in exponential back-off and rate limiting, and a nice
dashboard.

It should be noted that the existing /fetchversions/ handler was renamed
to /requeue/ and slightly modified: now it looks for versions that are
eligible for re-processing and reschedules them on the queue, which can
be useful if they have either timed-out of the queue or if the queue had
to be purged for some reason.

This change has been deployed and tested on AppEngine: there were
initially some failures due to memory limits, but those have thus far
disappeared after upgrading to the F4_1G instance class. Note that in
the passive state of no cron tasks we will scale down to zero instances,
and in the active state of many cron requests even the F4_1G instance
will scale up, so it is likely that increasing the instance size has
minimal impact on overall cost, though this has not been rigorously
investigated.

Other failures are rare and appear to be caused by deterministic
failure (e.g. no packages), or due to context timeout.

There were a few other internal changes that had to be made to enable
this CL. Notably:

 + The derrors package was refactored slightly to make it easier to wrap
 an error while preserving our internal error semantics, for example
 when you want to preserve a NotFound error from a client.

 + The conflict action was surfaced in the bulkInsert API, to allow for
 upserting new IndexVersions.

 + All use of the old VersionLogs type were removed, since it was easier
 to delete this functionality as part of this CL.

The net result is a rather large changeset, but one that is IMO easier
to review as a whole than to split up.

Fixes b/131614520
Fixes b/131612196
Fixes b/130540724
Updates b/131614050
Fixes b/131615422
Fixes b/132153576

Change-Id: I062995a213c99a660791608e25415b65cd7525a7
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/468418
Reviewed-by: Julie Qiu <julieqiu@google.com>