internal/postgres: search by parts of sub paths
Package path elements that are delimited by a dot and dash are now split
on the delimiter and indexed into documents.tsv_search_tokens for search.

For example, searching for "k8s" or "client" will now return
"k8s.io/go-client" (once the package is reindexed in the
documents table).

Updates b/131861451

Change-Id: I628ce5494ca589b8efb264a0be3b82cfcb1ed25c
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/470676
Reviewed-by: Robert Findley <rfindley@google.com>