internal/fetch: correctly parse all type of pseudo-versions
We were only handling the major-version form of pseudo-version described
at https://tip.golang.org/cmd/go/#hdr-Pseudo_versions. This CL adds
support for the prerelease and minor-version form. It also removes a
check that commit time is in the past, since the Go command appears not
to do this.

Fixes b/133792079

Change-Id: If551c8d178c29d569ca1e336a021e3f0e6f0c7eb
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/476536
Reviewed-by: Julie Qiu <julieqiu@google.com>