content,internal: add support for standard library packages
The Go standard library packages are now supported on the discovery
site, under the module "std".

It is also now possible to make fetch requests to the cron for module
versions using a tag or commit hash.

The versions template is also updated based on the current mock.

Updates b/131859847
Updates b/132875076
Updates b/126597197

Change-Id: I01fbbbff5bfec93d0ee4c5048f98ffe2da2f2a6a
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/472904
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>