internal/proxy: fix bug for encoding URL and add tests
A bug is fixed where requests to the proxy zip endpoint were not
encoded, which was introduced in commit b65b720.

Additional tests are added to internal/proxy to catch cases like this in
the future.

internal/proxy/test_helper.go is rearranged so that exported functions
are moved to the top of the file, but no logic is changed.

Change-Id: I3f956ed5b93667c4a909d7c2ac942cba8588c2dd
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/486063
Reviewed-by: Robert Findley <rfindley@google.com>