internal/proxy: return canonical version info request
The version returned by proxy.GetInfo is now always
the canonical version.

Previously, if a request for module "std@v.12" was made,
the version returned would be v1.12. Now it is v1.12.0.

The log for fetching a version is moved from handleFetch
to fetchAndInsertVersion, so that the canonical version is
logged instead of the requested version.

Fixes b/135267786

Change-Id: I17851697c6a65c6c332b69a37a18be50d96b9f20
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/489036
Reviewed-by: Robert Findley <rfindley@google.com>