content,internal/frontend: add HTTP error pages
frontend.ErrorPage and content/static/html/pages/error.tmpl are added.
These are used to generate error pages when an HTTP error occurs.

Errors are handled as follows:

* If a /<path>@<version> request returns a package that exists, but not
at the requested version, the user is shown a link to the versions view
for that package.

* If a /<path>@<version> request contains a version that is not a valid
semantic version, the user is informed that <version> is not a valid
semantic version, and shown a link to search for <path>/

* Otherwise, a user is shown an error page containing the status code
and text.

Additional minor changes:

* The check for semver.IsValid is moved out of
frontend.parseModulePathAndVersion, to allow for the feature of
informing users of invalid semantic version requests described above.

* The @{{.Version}} is removed from paths in search.tmpl, in case
mvw_search_documents is out of sync of versions.

Fixes b/132195669
Fixes b/132929093
Fixes b/129755686
Fixes b/130190598
Fixes b/132875729

Change-Id: I06f85795e79e6a90a6ed48401f4197b5a8e866b3
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/488380
Reviewed-by: Robert Findley <rfindley@google.com>