discovery: add opencensus instrumentation
Opencensus instrumentation is added via the ochttp and ocsql
integrations. Additionally, debug pages are served on a separate port to
view the opencensus zpage for trace spans and the prometheus page for
metrics.

If a GCP project is detected, this debug information is exported to
Stackdriver.

After testing, the dev queue is significantly scaled-up, to accomodate
the newly optimized fetch.

Updates b/132629518

Change-Id: I7b0aa96e46c30568d320c3820ba55d676f39590b
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/493294
Reviewed-by: Julie Qiu <julieqiu@google.com>