internal/frontend: bypass search results only if query contains "/"
A search query will jump directly to the package details page and bypass
search results if it passes a package path in our database. However,
when a user searches for "context" or "errors", they might not
necessarily be looking for these stdlib packages.

The search query will now only bypass search results if the query
contains a "/".

Fixes b/135765305

Change-Id: Ice2e5b5fdb5bec17480f5612675acd9f7e68d160
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/494910
Reviewed-by: Robert Findley <rfindley@google.com>