internal/frontend: serve search results from /search to avoid a redirect
When given a path, if http.ServeMux cannot match it directly and it
does not have a trailing slash, it will attempt to match on the path
with a trailing slash and redirect if it succeeds. This behavior was
causing about an 80ms hit to search results page latency.

We also don’t want to be serving /search/ since a user could request
/search/garbagestuff?q=foo and it would serve the search page, busting
any caching in place.

Change-Id: I8734e1186c3d919177403bb9fce610e9bbc37023
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/497398
Reviewed-by: Julie Qiu <julieqiu@google.com>