content/static: restyle package tabs to match designs
Adds a horizontal scrolling behavior on mobile and some minor JS
to scroll the selected tab into view if it happens to be off screen.

Additionally allow the search form input to shrink to accomodate
smaller screens. The placeholder text is cut off on much smaller
screens but that will be addressed in a follow-up change.

Fixes b/131913479
Fixes b/135532582

Change-Id: I3ee4496a33886f184e61eeabc835c5a9e96bafdc
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/497406
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>