internal/frontend,content: refactor fetchVersionDetails
fetchVersionDetails is refactored to use a new versionTree type, which
exposes functionality for traversing the version hierarchy. There seemed
to be enough version-related functions to warrant a new version.go file
within the frontend package.

As a result of this refactoring, the current module is sorted to the
top.

Fixes b/132875076
Fixes b/136683375

Change-Id: I432f66df94018fc75a82d9b1ba2e01c47da20912
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/501696
Reviewed-by: Julie Qiu <julieqiu@google.com>