internal/etl: use file name, not path, in Context.MatchFile call
The file name parameter (second parameter) to Context.MatchFile call
needs to be just the file name. It's not okay to give a longer path,
because its prefix and inner contents are checked as part of the build
constraint satisfaction logic, not just the suffix.

More specifically, many .go file paths had dots in the full file path.
E.g., some/module@v1.0.0/pkg/file.go. The dots in v1.0.0 were being
matched as the file extension dot, causing MatchFile to report
incorrect results.

Document this requirement in the documentation of the matchFile helper.

Add a test case to prevent regressions.

Fixes b/136813944

Change-Id: Ic2a79b5b076be18aefcc1d7ddc2129459e1d0f66
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/504897
Reviewed-by: Robert Findley <rfindley@google.com>