content: implement text styleguide
This changes corrects the text style in a handful of places. I verified
the base font-size with our designer (16pt, 16px on screen, or 1rem).
This change also changes all font-size units from px to rem, which is in
line with our accessibility guide.

A small number of styles are removed, as they were either redundant or
were replaced by the more correct line-height.

Updates b/135946866

Change-Id: Ibfcab2d5d7afab20f9fbc0653822d49f8ce77f68
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/505097
Reviewed-by: Julie Qiu <julieqiu@google.com>