internal/etl: change derrors.InvalidArgument to terminal error
Previously, requests from the proxy that returned with a status 400 were
inserted with status 500 into the module_version_states table.

As a result, a module version like
"github.com/youzan/nsq/@v/v0.3.7-HA.1.9.4.1" would be retried
indefinitely.

These bad requests are now registered as status 400 in the
module_version_states table so that it is a terminal error.

Change-Id: I3fd8026068939fdedfa2fde92b1d838dfc0e7182
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/506484
Reviewed-by: Andrew Bonventre <andybons@google.com>