internal/etl/internal/doc: add copy of go/doc with support for classifying Examples
This change adds a temporary copy of the go/doc package from the
standard library with the CL 94855 (golang.org/cl/94855) by Joe Tsai
applied. That CL is a work in progress on adding support for classifying
Examples, an important task that every godoc implementation needs, which
takes a non-trivial amount of logic to implement.

The motivation here is to copy go/doc and evaluate and iterate on that
CL in the context of discovery documentation rendering needs. It's much
easier to evaluate it when using it concretely. After the implementation
and API become stable and the tree unfreezes after Go 1.13 is out, the
changes here can be upstreamed into CL 94855 and merged into go/doc.

The code here is mostly go/doc from Go 1.12, with CL 94855 by Joe Tsai
applied and rebased on top (with merge conflicts resolved), and some
code changes on top by me. It's functional as is, but there will be
more changes in the future.

Updates b/137567588

Change-Id: Iaec5fdea49a107e54e1c27425908b690129f9ac8
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/504148
Reviewed-by: Julie Qiu <julieqiu@google.com>