internal/etl/dochtml/internal/render: add copy of godoc render package from CL 72890
This change adds a temporary copy of a new godoc/internal/render package
from CL 72890 (golang.org/cl/72890). That CL is a work in progress
created by Joe Tsai that implements more godoc rendering, incorporating
more modern practices and with better security properties.

The motivation here is to copy that package and evaluate and iterate on
that CL in the context of discovery documentation rendering needs. It's
easier to evaluate it when using it concretely. After the implementation
and API become stable, we will think about where to place the final code
(with a non-internal import path) and make that available.

The code here is mostly from CL 72890 by Joe Tsai, with code changes on
top of by me. It has been updated to use the go/doc Example classification
code in the previous CL. It's functional as is, but there will be
more changes in the future.

Updates b/137567588

Change-Id: Ibcd982c8a6ff9aee16920487d64ce21a539e2f92
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/504149
Reviewed-by: Julie Qiu <julieqiu@google.com>